Add DebugCopying GC
This GC uses malloc to allocate objects, but otherwise behaves like a semi-space GC.

This hopefully enables us to use memory/address senatizers to catch GC bugs.

+++ src/memory/MarkSweepHeap.h
@@ -12,6 +12,6 @@ public:
 private:
     vector<AbstractVMObject*>* allocatedObjects;
     size_t spcAlloc;
-    long collectionLimit;
+    size_t collectionLimit;
 
 };
